code factorization
