+++ src/data_ingestion.py
@@ -2,19 +2,7 @@ import pandas as pd
 import yaml
 import os
 from sklearn.model_selection import train_test_split
-
-def load_data(input_file_path):
-    try:
-        df = pd.read_csv(input_file_path)
-        print(f"Data loaded successfully from {input_file_path}")
-        return df
-    except FileNotFoundError:
-        print(f"Error: File not found at path '{input_file_path}'")
-    except pd.errors.ParserError:
-        print("Error: Failed to parse CSV file.")
-    except Exception as e:
-        print(f"Unexpected error while loading data: {e}")
-        return None
+from common.utils import load_data
 
 def split_data(df, output_path, test_ratio, random_state):
     try:
